frontend/requirements.txt
# local package
-e .

# frontend requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
streamlit
streamlit-extras
pyyaml
pytest